requirements.txt
# Core dependencies
fastapi
uvicorn
pydantic
python-dotenv
httpx
asyncio
aiohttp
beautifulsoup4
lxml
brotli

# Vector database and embeddings
pinecone
openai
langchain
langchain-pinecone
langchain-openai
langchain-text-splitters

# Agent framework
langchain-core
langgraph
anthropic

# API and UI
mcp

# Utilities
loguru
tqdm
tenacity
pyyaml
markitdown
langchain-text-splitters
